+++ sandbox/tests/collision/passive.ts
@@ -4,7 +4,7 @@ var game = new ex.Engine({
   height: 400
 });
 
-ex.Physics.collisionResolutionStrategy = ex.CollisionResolutionStrategy.Arcade;
+ex.Physics.collisionResolutionStrategy = ex.SolverStrategy.Arcade;
 
 var activeBlock = new ex.Actor({x: 200, y: 200, width: 50, height: 50, color: ex.Color.Red.clone()});
 activeBlock.body.collisionType = ex.CollisionType.Active;