+++ src/spec/ArcadeSolverSpec.ts
@@ -1,6 +1,7 @@
 import * as ex from '@excalibur';
 import { ExcaliburMatchers } from 'excalibur-jasmine';
 import { TestUtils } from './util/TestUtils';
+import { DefaultPhysicsConfig } from '../engine/Collision/PhysicsConfig';
 
 describe('An ArcadeSolver', () => {
   beforeAll(() => {
@@ -24,7 +25,7 @@ describe('An ArcadeSolver', () => {
     const contacts = [...pair1.collide(), ...pair2.collide()];
     expect(contacts.length).toBe(2);
 
-    const sut = new ex.ArcadeSolver();
+    const sut = new ex.ArcadeSolver(DefaultPhysicsConfig.arcade);
 
     for (const contact of contacts) {
       sut.solvePosition(contact);
@@ -91,7 +92,7 @@ describe('An ArcadeSolver', () => {
   });
 
   it('should cancel collision contacts where there is no more overlap', () => {
-    const arcadeSolver = new ex.ArcadeSolver();
+    const arcadeSolver = new ex.ArcadeSolver(DefaultPhysicsConfig.arcade);
 
     const player = new ex.Actor({
       x: 0,
@@ -127,7 +128,7 @@ describe('An ArcadeSolver', () => {
 
   it('should NOT cancel collisions where the bodies are moving away from the contact', () => {
 
-    const arcadeSolver = new ex.ArcadeSolver();
+    const arcadeSolver = new ex.ArcadeSolver(DefaultPhysicsConfig.arcade);
 
     const player = new ex.Actor({
       x: 0,
@@ -166,7 +167,7 @@ describe('An ArcadeSolver', () => {
   });
 
   it('should cancel near zero mtv collisions', () => {
-    const arcadeSolver = new ex.ArcadeSolver();
+    const arcadeSolver = new ex.ArcadeSolver(DefaultPhysicsConfig.arcade);
 
     const player = new ex.Actor({
       x: 0,
@@ -203,7 +204,7 @@ describe('An ArcadeSolver', () => {
   });
 
   it('should cancel near zero overlap collisions', () => {
-    const arcadeSolver = new ex.ArcadeSolver();
+    const arcadeSolver = new ex.ArcadeSolver(DefaultPhysicsConfig.arcade);
 
     const player = new ex.Actor({
       x: 0,
@@ -232,7 +233,7 @@ describe('An ArcadeSolver', () => {
   });
 
   it('should cancel zero overlap collisions during presolve', () => {
-    const arcadeSolver = new ex.ArcadeSolver();
+    const arcadeSolver = new ex.ArcadeSolver(DefaultPhysicsConfig.arcade);
 
     const player = new ex.Actor({
       x: 0,
@@ -260,5 +261,63 @@ describe('An ArcadeSolver', () => {
     arcadeSolver.preSolve([contact]);
     // Considers infinitesimally overlapping to no longer be overlapping and thus cancels the contact
     expect(contact.isCanceled()).toBe(true);
+  });
+
+  it('should allow solver bias and solve certain contacts first', async () => {
+    const game = TestUtils.engine({
+      width: 1000,
+      height: 1000,
+      fixedUpdateFps: 60,
+      physics: {
+        gravity: ex.vec(0, 5000),
+        solver: ex.SolverStrategy.Arcade,
+        arcade: {
+          contactSolveBias: ex.ContactSolveBias.VerticalFirst
+        }
+      }
+    });
+    const clock = game.clock as ex.TestClock;
+    await TestUtils.runToReady(game);
+    // big tiles so distance heuristic doesn't work
+    const lastPos = ex.vec(0, 0);
+    for (let x = 0; x < 10; x++) {
+      const width = (x % 2 === 1 ? 16 : 200);
+      game.add(
+        new ex.Actor({
+          name: 'floor-tile',
+          x: lastPos.x,
+          y: 300,
+          width: width,
+          height: x % 2 ? 16 : 900,
+          anchor: ex.Vector.Zero,
+          color: ex.Color.Red,
+          collisionType: ex.CollisionType.Fixed
+        })
+      );
+      lastPos.x += width;
+    }
+
+    const player = new ex.Actor({
+      pos: ex.vec(100, 270),
+      width: 16,
+      height: 16,
+      collisionType: ex.CollisionType.Active,
+      color: ex.Color.Red
+    });
+
+    // place player on tiles
+    player.vel.x = 164;
+    game.add(player);
+
+    // run simulation and ensure now left/right contacts are generated
+    player.on('postcollision', evt => {
+      expect(evt.side).not.toBe(ex.Side.Left);
+      expect(evt.side).not.toBe(ex.Side.Right);
+      expect(evt.side).toBe(ex.Side.Bottom);
+    });
+
+    for (let i = 0; i < 40; i++) {
+      clock.step(16);
+    }
   });
 });