+++ src/spec/CollisionContactSpec.ts
@@ -2,6 +2,7 @@ import * as ex from '@excalibur';
 import { TransformComponent } from '@excalibur';
 import { EulerIntegrator } from '../engine/Collision/Integrator';
 import { MotionComponent } from '../engine/EntityComponentSystem/Components/MotionComponent';
+import { DefaultPhysicsConfig } from '../engine/Collision/PhysicsConfig';
 
 describe('A CollisionContact', () => {
   let actorA: ex.Actor;
@@ -58,7 +59,7 @@ describe('A CollisionContact', () => {
       null
     );
 
-    const solver = new ex.ArcadeSolver();
+    const solver = new ex.ArcadeSolver(DefaultPhysicsConfig.arcade);
     solver.solve([cc]);
 
     expect(actorA.pos.x).toBe(-0.5);
@@ -88,7 +89,7 @@ describe('A CollisionContact', () => {
       null
     );
 
-    const solver = new ex.ArcadeSolver();
+    const solver = new ex.ArcadeSolver(DefaultPhysicsConfig.arcade);
     solver.solve([cc]);
 
     expect(actorAPreCollide).toHaveBeenCalledTimes(1);
@@ -119,9 +120,8 @@ describe('A CollisionContact', () => {
       [new ex.Vector(10, 0)],
       null
     );
-    ex.Physics.slop = 0; // slop is normally 1 pixel, we are testing at a pixel scale here
-    const solver = new ex.RealisticSolver();
-
+    // slop is normally 1 pixel, we are testing at a pixel scale here
+    const solver = new ex.RealisticSolver({...DefaultPhysicsConfig.realistic, slop: 0});
     // Realistic solver converges over time
     for (let i = 0; i < 4; i++) {
       solver.solve([cc]);
@@ -166,8 +166,8 @@ describe('A CollisionContact', () => {
       [new ex.Vector(10, 0)],
       null
     );
-    ex.Physics.slop = 0; // slop is normally 1 pixel, we are testing at a pixel scale here
-    const solver = new ex.RealisticSolver();
+    // slop is normally 1 pixel, we are testing at a pixel scale here
+    const solver = new ex.RealisticSolver({ ...DefaultPhysicsConfig.realistic, slop: 0});
 
     solver.solve([cc]);
     // Realistic solver uses velocity impulses to correct overlap
@@ -211,8 +211,7 @@ describe('A CollisionContact', () => {
       [new ex.Vector(10, 0), new ex.Vector(10, 10)],
       null
     );
-    ex.Physics.slop = 0; // slop is normally 1 pixel, we are testing at a pixel scale here
-    const solver = new ex.RealisticSolver();
+    const solver = new ex.RealisticSolver(DefaultPhysicsConfig.realistic);
     solver.solve([cc]);
     // Realistic solver uses velocity impulses to correct overlap
     EulerIntegrator.integrate(actorA.get(TransformComponent), actorA.get(MotionComponent), ex.Vector.Zero, 30);
@@ -250,7 +249,10 @@ describe('A CollisionContact', () => {
       null
     );
     ex.Physics.slop = 0; // slop is normally 1 pixel, we are testing at a pixel scale here
-    const solver = new ex.RealisticSolver();
+    const solver = new ex.RealisticSolver({
+      ...DefaultPhysicsConfig.realistic,
+      slop: 0
+    });
     solver.solve([cc]);
     // Realistic solver uses velocity impulses to correct overlap
     EulerIntegrator.integrate(actorA.get(TransformComponent), actorA.get(MotionComponent), ex.Vector.Zero, 30);
@@ -287,7 +289,7 @@ describe('A CollisionContact', () => {
       null
     );
     ex.Physics.slop = 0; // slop is normally 1 pixel, we are testing at a pixel scale here
-    const solver = new ex.RealisticSolver();
+    const solver = new ex.RealisticSolver(DefaultPhysicsConfig.realistic);
     solver.solve([cc]);
     // Realistic solver uses velocity impulses to correct overlap
     EulerIntegrator.integrate(actorA.get(TransformComponent), actorA.get(MotionComponent), ex.Vector.Zero, 30);
@@ -323,7 +325,7 @@ describe('A CollisionContact', () => {
     );
 
 
-    const solver = new ex.RealisticSolver();
+    const solver = new ex.RealisticSolver(DefaultPhysicsConfig.realistic);
     solver.solve([cc]);
 
     expect(emittedA).toBe(true);